Changed file structure, unified documentation, implemented dynamic list loader.
